split tests into different logical sections
