+++ pyproject.toml
@@ -7,15 +7,15 @@ readme = "README.md"
 
 [tool.poetry.dependencies]
 python = ">=3.10,<3.13"
-crewai = "0.19.0"
 crewai-tools = "0.0.15"
 pyright = "^1.1.350"
 click = "^8.1.7"
-python-dotenv = "^1.0.1"
 safe-eth-py = "^5.8.0"
 eth-account = "^0.11.0"
 web3 = "^6.15.1"
 requests = "^2.31.0"
+python-dotenv = "^1.0.0"
+crewai = "0.22.5"
 
 [tool.poetry.group.dev.dependencies]
 mypy = "^1.8.0"
